requirements.txt
streamlit
langchain
streamlit-option-menu
PyMuPDF
flask
python-dotenv
pinecone-client
openai
tiktoken